fix linting other issues
