+++ .github/workflows/pr.yaml
@@ -390,7 +390,7 @@ jobs:
           # Use release profile in CI for faster builds
           cargo build --release --bin gateway --bin mock-inference-provider
           
-      - name: Start Redis for rate limiting test
+      - name: Start Redis for performance test
         run: |
           docker compose -f tensorzero-internal/tests/e2e/docker-compose.yml up -d redis
           # Wait for Redis to be ready
@@ -399,39 +399,22 @@ jobs:
           
       - name: Run performance test
         run: |
-          ./ci/performance/run-performance-test-ci.sh
-          
-      - name: Run performance test with rate limiting
-        run: |
           # Export Redis URL for rate limiting
           export TENSORZERO_REDIS_URL="redis://localhost:6379"
           
-          # Copy proper config with rate limiting enabled
-          cp ci/performance/tensorzero-perf-test-proper-rate-limit.toml ci/performance/tensorzero-perf-test.toml.tmp
-          mv ci/performance/tensorzero-perf-test.toml ci/performance/tensorzero-perf-test.toml.bak
-          mv ci/performance/tensorzero-perf-test.toml.tmp ci/performance/tensorzero-perf-test.toml
-          
-          # Run test
+          # Run performance test (rate limiting is enabled by default in config)
           ./ci/performance/run-performance-test-ci.sh
-          mv performance-results.json performance-results-with-rate-limit.json
           
-          # Restore original config
-          mv ci/performance/tensorzero-perf-test.toml.bak ci/performance/tensorzero-perf-test.toml
-          
-          # Compare results
-          echo "━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━"
-          echo "Rate Limiting Performance Impact:"
-          P99_BASE=$(jq -r '.latencies."99th"' performance-results.json | awk '{print $1/1000000}')
-          P99_RL=$(jq -r '.latencies."99th"' performance-results-with-rate-limit.json | awk '{print $1/1000000}')
-          echo "P99 without rate limiting: ${P99_BASE}ms"
-          echo "P99 with rate limiting: ${P99_RL}ms"
-          OVERHEAD=$(awk -v base="$P99_BASE" -v rl="$P99_RL" 'BEGIN { printf "%.2f", ((rl - base) / base) * 100 }')
-          echo "Overhead: ${OVERHEAD}%"
+          # Check P99 latency threshold (rate limiting should keep overhead minimal)
+          P99=$(jq -r '.latencies."99th"' performance-results.json | awk '{print $1/1000000}')
+          echo "P99 latency with rate limiting: ${P99}ms"
           
           # Ensure rate limiting doesn't exceed threshold
-          if [ $(awk -v p99="$P99_RL" 'BEGIN { print (p99 > 1.5) }') -eq 1 ]; then
+          if [ $(awk -v p99="$P99" 'BEGIN { print (p99 > 1.5) }') -eq 1 ]; then
               echo "❌ Error: P99 latency with rate limiting exceeds 1.5ms threshold"
               exit 1
+          else
+              echo "✅ P99 latency with rate limiting is within acceptable limits"
           fi
           
       - name: Upload performance results
