+++ ci/performance/tensorzero-perf-test.toml
@@ -1,13 +1,29 @@
-# Performance test configuration with authentication disabled
+# Performance test configuration with rate limiting enabled
 
 [gateway]
 authentication.enabled = true
 observability.enabled = false
 
+# Rate limiting configuration
+[gateway.rate_limits]
+enabled = true
+
 # Models
 [models."gpt-3.5-turbo"]
 routing = ["openai"]
 endpoints = ["chat"]
+
+# Performance-optimized rate limits
+[models."gpt-3.5-turbo".rate_limits]
+algorithm = "fixed_window"
+requests_per_second = 1000    # High limit for performance testing
+burst_size = 2000             # Allow bursts
+enabled = true
+
+# Performance settings - optimized for low latency
+cache_ttl_ms = 200           # 200ms cache
+local_allowance = 0.8        # 80% local decisions
+sync_interval_ms = 100       # Sync every 100ms
 
 [models."gpt-3.5-turbo".providers.openai]
 type = "openai"
